libcamera: buffer: Improve timestamp documentation
The Buffer timestamp documentation is short and does not explain that
the timestamp is based on a monotonic time source, as opposed to the
real/wall-time clocks available in the system.

Expand upon the statements to detail that it can not be used directly as
a wall-clock for the captured date/time without a further reference
point, and that by taking a reference point - the user will be bringing
in potential inconsistencies which must be considered.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Umang Jain <email@uajain.com>