libcamera: request: Make it extensible
Provide an extensible private object for the Request class.
This allows us to make modifications to the private API and storage of
requests without affecting the public API or ABI.

The D pointer is obtained in all Request functions implemented in the
request.cpp file along with an assertion that the D pointer was valid to
provide extra validation checking that the Request has not been
deleted while in use, as it is 'owned' by the application.

Reviewed-by: Jean-Michel Hautbois <jeanmichel.hautbois@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>

---
The assertions added in findBuffer, complete() and completeBuffer()
allow us to ensure that the Request is still valid while asynchronous
actions occur on the Request internally in libcamera, and provide
(almost) equivalent functionality as the Request Canary previously
proposed.

The additions in reuse() and addBuffer() are called from applications,
so the assertions may be less helpful there, but I've added them for
consistency.