utils: ipc: Support types defined in core.mojom with skipHeader only
Previously, types that were defined in core.mojom with skipHeader only
and no skipSerdes would cause serializer generation to fail due to the
generated IPA namespace being used for mojom-defined types instead of
the libcamera namespace. This doesn't work because skipHeader means that
the type is defined in the libcamera (or some other) namespace
elsewhere.

Fix this by using "using namespace" in the generated serializer header,
so that mojom-defined IPA types can be used without the full namespace
qualifier.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>