meson: summarise paths
The IPAs are loaded from a defined location set as IPA_MODULE_DIR.
Further more, configuration files for both libcamera and IPAs are loaded
from paths defined during the meson configuration.

To make it easy to identify these locations for a given configuration,
present them as a 'paths' section in the meson summary.

This adds a summary section to meson in the form:

  Paths
    LIBCAMERA_DATA_DIR       : /usr/share/libcamera
    LIBCAMERA_SYSCONF_DIR    : /etc/libcamera
    IPA_PROXY_DIR            : /usr/libexec/libcamera
    IPA_CONFIG_DIR           : /etc/libcamera/ipa:/usr/share/libcamera/ipa
    IPA_MODULE_DIR           : /usr/lib/x86_64-linux-gnu/libcamera

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>